removes #do_not_eager_load
There is in the horizon a potential speed up in VM disabling tracing:

    https://bugs.ruby-lang.org/issues/14104

While nowadays we have measured no performance penalty by the existence
of a trace point on the :class event (to support explicit namespaces),
if that plan goes on, one single gem dependency having eager load
exceptions could leave a trace point active, therefore preventing the
optimization for the entire process.

Want to thank @SamSaffron for bringing that issue to may attention.

There is no 1.0 release of Zeitwerk yet, I am on time to remove this and
anticipate to that potential problem. If things do not go that way in
the end, we can always revisit.

As @josevalim says, adding is cheap, removing is hard.