rewrite set_autoloads_in_namespace as a callback
The method name set_autoloads_in_namespace clashes a bit with the
existing set_autoloads_in_dir. The concepts "dir" and "namespace" are
close, but they are different in the library nomenclature, so how is it
possible that you can set autoloads for both things? There's a bit of
overlap that in my mind generates a bit of confusion.

Also, you don't always want to set autoloads in a namespace. Visiting a
directory is more intentional, while normally you'll have a good number
of classes and modules for which there are no directories to visit. Both
in the managed directories, and outside, since the tracer is called when
_any_ class or module is created or reopened in the entire process.

On the other hand, it could be argued that the tracer is a code
conceptually far away from Zeitwerk's logic, like Kernel#require is.

All consider, I prefer to treat this code as a callback.