optimize shadow files tracking [closes #36]
The purpose of this collection is twofold: on one hand skip shadowed
files while eager loading, and on the other hand storing some related
information for debugging.

The first use case is telling us shadowed files are not needed after
eager loading, therefore we can clear them.

The second use case does not match the current design principle in which
troubleshooting is done via logging, not introspection. This principle
agrees with experience using Z, and allows us to save RAM.