inflect with capitalize [closes #59]
While the default inflector is extremely delicate and considered to be
frozen, capitalize! is there for performance. Problem with the bang
version is that it returns nil for strings that start with a number, for
example, resulting in the potential deletion of segments. See the test
in the patch for a use case. I believe this behavior is wrong.

I do not think this is going to break existing code, since such
constants are not usable.